Fix circular dependency PetabImporter-PetabAmiciObjective; tidy up and extend docs (#274)
* allow arbitrary args for petab amici

* fix circular dependence PetabImporter-PetabAmiciObjective; tidy up and extend docs

* separate amici objective and utils

* tidy up

* move further funs to util

* minor edit
