+++ doc/api_logging.rst
@@ -1,0 +1,5 @@
+.. automodule:: pypesto.logging
+   :members:
+   :inherited-members:
+   :special-members:
+   :imported-members:
