+++ doc/api_petab.rst
@@ -1,0 +1,5 @@
+.. automodule:: pypesto.petab
+   :members:
+   :inherited-members:
+   :special-members:
+   :imported-members:
