+++ doc/api_storage.rst
@@ -1,0 +1,5 @@
+.. automodule:: pypesto.storage
+   :members:
+   :inherited-members:
+   :special-members:
+   :imported-members:
