+++ pypesto/objective/__init__.py
@@ -2,10 +2,11 @@
 Objective
 =========
 
+The :class:`pypesto.Objective` defines the objective function.
 """
 
 from .objective import Objective
-from .amici_objective import AmiciObjective
+from .amici_objective import AmiciObjective, AmiciObjectBuilder
 from .aggregated import AggregatedObjective
 from .util import res_to_chi2, sres_to_schi2
 from .history import (
@@ -17,9 +18,3 @@ from .history import (
     Hdf5History,
     OptimizerHistory)
 from . import constants
-
-# PEtab is an optional dependency
-try:
-    from .petab_import import PetabImporter
-except ModuleNotFoundError:
-    PetabImporter = None
