+++ pypesto/problem.py
@@ -1,6 +1,6 @@
 """
 Problem
--------
+=======
 
 A problem contains the objective as well as all information like prior
 describing the problem to be solved.
