+++ pypesto/profile/__init__.py
@@ -1,7 +1,6 @@
 """
 Profile
 =======
-
 """
 
 from .profile import (
