+++ pypesto/result.py
@@ -1,6 +1,6 @@
 """
 Result
-------
+======
 
 The pypesto.Result object contains all results generated by
 the pypesto components. It contains sub-results for
